+++ src/Steps/Loading/Http/Paginator.php
@@ -3,6 +3,7 @@
 namespace Crwlr\Crawler\Steps\Loading\Http;
 
 use Crwlr\Crawler\Steps\Html\DomQueryInterface;
+use Crwlr\Crawler\Steps\Loading\Http\Paginators\QueryParamsPaginator;
 use Crwlr\Crawler\Steps\Loading\Http\Paginators\SimpleWebsitePaginator;
 
 class Paginator
@@ -14,5 +15,10 @@ class Paginator
         int $maxPages = self::MAX_PAGES_DEFAULT,
     ): SimpleWebsitePaginator {
         return new SimpleWebsitePaginator($paginationLinksSelector, $maxPages);
+    }
+
+    public static function queryParams(int $maxPages = Paginator::MAX_PAGES_DEFAULT): QueryParamsPaginator
+    {
+        return new QueryParamsPaginator($maxPages);
     }
 }