+++ src/Steps/Loading/Http/Paginators/QueryParams/QueryParamManipulator.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Crwlr\Crawler\Steps\Loading\Http\Paginators\QueryParams;
+
+use Crwlr\QueryString\Query;
+
+interface QueryParamManipulator
+{
+    public function execute(Query $query): Query;
+}