+++ tests/_Integration/Server.php
@@ -33,6 +33,10 @@ if (str_starts_with($route, '/paginated-listing')) {
     return include(__DIR__ . '/_Server/PaginatedListing.php');
 }
 
+if (str_starts_with($route, '/query-param-pagination')) {
+    return include(__DIR__ . '/_Server/QueryParamPagination.php');
+}
+
 if ($route === '/blog-post-with-json-ld') {
     return include(__DIR__ . '/_Server/BlogPostWithJsonLd.php');
 }